Create separate file test_freestyle_project_configuration.py PR#625
